[feat] song error setting

+++ GIS_server/src/main/java/org/pyriboo/gis_server/domain/song/service/SongServiceImpl.java
@@ -1,7 +1,6 @@
 package org.pyriboo.gis_server.domain.song.service;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.pyriboo.gis_server.domain.playlist.dto.AddSongReq;
@@ -11,6 +10,8 @@ import org.pyriboo.gis_server.domain.song.dto.SongReq;
 import org.pyriboo.gis_server.domain.song.dto.SongRes;
 import org.pyriboo.gis_server.domain.song.model.Song;
 import org.pyriboo.gis_server.domain.song.repository.SongRepository;
+import org.pyriboo.gis_server.global.error.exception.SongException;
+import org.pyriboo.gis_server.global.error.type.SongErrorType;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -25,13 +26,16 @@ public class SongServiceImpl implements SongService {
 
 	@Override
 	public SongRes createSong(SongReq songReq) {
-		Song song = new Song();
-		song.setTitle(songReq.getTitle());
-		song.setArtist(songReq.getArtist());
-		song.setHashtags(songReq.getHashtags());
-		Song savedSong = songRepository.save(song);
-
-		return convertToDto(savedSong);
+		try {
+			Song song = new Song();
+			song.setTitle(songReq.getTitle());
+			song.setArtist(songReq.getArtist());
+			song.setHashtags(songReq.getHashtags());
+			Song savedSong = songRepository.save(song);
+			return convertToDto(savedSong);
+		} catch (Exception e) {
+			throw new SongException(SongErrorType.SONG_CREATION_FAILED);
+		}
 	}
 
 	@Override
@@ -43,36 +47,43 @@ public class SongServiceImpl implements SongService {
 
 	@Override
 	public SongRes getSongById(String id) {
-		Optional<Song> song = songRepository.findById(id);
-		return song.map(this::convertToDto).orElse(null);
+		Song song = songRepository.findById(id)
+			.orElseThrow(() -> new SongException(SongErrorType.SONG_NOT_FOUND));
+		return convertToDto(song);
 	}
 
 	@Override
 	public SongRes updateSong(String id, SongReq songReq) {
-		Optional<Song> existingSong = songRepository.findById(id);
-		if (existingSong.isPresent()) {
-			Song song = existingSong.get();
-			song.setTitle(songReq.getTitle());
-			song.setArtist(songReq.getArtist());
-			song.setHashtags(songReq.getHashtags());
-
-			Song updatedSong = songRepository.save(song);
+		Song existingSong = songRepository.findById(id)
+			.orElseThrow(() -> new SongException(SongErrorType.SONG_NOT_FOUND));
+		try {
+			existingSong.setTitle(songReq.getTitle());
+			existingSong.setArtist(songReq.getArtist());
+			existingSong.setHashtags(songReq.getHashtags());
+			Song updatedSong = songRepository.save(existingSong);
 			return convertToDto(updatedSong);
+		} catch (Exception e) {
+			throw new SongException(SongErrorType.SONG_UPDATE_FAILED);
 		}
-		return null;
 	}
 
 	@Override
 	public void addSongToPlaylist(AddSongReq addSongReq) {
 		Playlist playlist = playlistRepository.findById(addSongReq.getPlaylistId())
-			.orElseThrow(() -> new RuntimeException("Playlist not found"));
+			.orElseThrow(() -> new SongException(SongErrorType.SONG_NOT_FOUND));
 		List<Song> songsToAdd = songRepository.findAllById(addSongReq.getSongIds());
+		if (songsToAdd.isEmpty()) {
+			throw new SongException(SongErrorType.SONG_NOT_FOUND);
+		}
 		playlist.getSongs().addAll(songsToAdd);
 		playlistRepository.save(playlist);
 	}
 
 	@Override
 	public void deleteSong(String id) {
+		if (!songRepository.existsById(id)) {
+			throw new SongException(SongErrorType.SONG_DELETION_FAILED);
+		}
 		songRepository.deleteById(id);
 	}
 
@@ -82,7 +93,6 @@ public class SongServiceImpl implements SongService {
 		songRes.setTitle(song.getTitle());
 		songRes.setArtist(song.getArtist());
 		songRes.setHashtags(song.getHashtags());
-
 		return songRes;
 	}
 }
